requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0

# LangGraph workflow
langgraph>=0.2.0

# Vector database
chromadb>=0.5.0

# Embeddings (local, no Ollama needed)
sentence-transformers>=3.0.0

# LLM via Groq API
httpx>=0.27.0

# MCP protocol
mcp>=1.0.0

# Web interface
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
